requirements.txt
requests>=2.31.0
urllib3<2.0.0
beautifulsoup4>=4.12.2
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
